refactor(voice): isolate agent task runtime
